chore: use polyglot library from cozy-template (and add date-fns) :globe_with_meridians:
