+++ app/components/DataItem.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { translate } from '../plugins/preact-polyglot'
+import { translate } from '../plugins/i18n'
 
 const DataItem = ({ t, dataType, hex }) => (
   <li className='account-data'>
