+++ app/components/UseCaseList.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { translate } from '../plugins/preact-polyglot'
+import { translate } from '../plugins/i18n'
 import UseCaseItem from './ConnectorItem'
 
 // Fallback to get the item background image and avoid error if not found
