+++ package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "csswring": "^5.1.1",
+    "date-fns": "^1.28.4",
     "node-polyglot": "2.0.0",
     "preact": "^8.1.0",
     "preact-compat": "^3.16.0",
